Rename store folder to state
