+++ app/index.js
@@ -10,7 +10,7 @@ import Immutable from 'seamless-immutable';
 
 import getRoutes from 'app/routes';
 import rootReducer from 'reducers/index';
-import configureStore from 'store/configureStore';
+import configureStore from 'state/configureStore';
 
 import 'assets/styles/app.less';
 import 'assets/styles/customization/espoo/customization.less';
